fix for fetchSDA() when multiple ecosites linked to 1 component
Currently fetchSDA() makes it easy for folks to query data by ecosite.
This fix maintains that feature, while warning the user when multiple
ecosites are linked to 1 component. The fix accommodates the issue by
concatinating the ecosites with a comma. Some components have this
issue, particularly if ecological sites weren't originally taken into
consideration when the map unit was designed.