stub related to soil data aggregation best practicies / raster data / deratify in the context of SPC
This is a rough start, but functional. I'm not convinced that
aggregation should happen in R vs. SQL, lilely both are needed. I'm
pushing our developers hard to implement SDA macros to simplify the SQL.
Until that time the code developed by @cferguso and Jason for SOD is a
fine start. There is also the code the Steve Peaslee has written in
python.